* fix, network.h: wrong type 2 32-bit ASN Route Distinguisher definition amended.

+++ src/network.h
@@ -265,7 +265,7 @@ struct rd_as4
 {
   u_int16_t type;
   as_t as;
-  u_int32_t val;
+  u_int16_t val;
 } __attribute__ ((packed));
 
 /* Picking one of the three structures as rd_t for simplicity */
